Move error messages in form, add links for registering on login page and vice versa for easy navigation
Error messages make more sense below the input instead of above the input. Follow the natural flow of progress in a form.
I also added links in LoginForm.tsx and RegisterForm.tsx.
In LoginForm.tsx, a link to the register page was added in case a user needs to register first.
In RegisterForm.tsx, a link to the login page was added in case a user is already registered and just wants to login.