Restructure Products page in a way that is easy on the eyes
Made images smaller.
Category title has a thin font size now, to separate it from the product title.
Prices are bold.
Product titles are full length now.
Div with product title, category, and price is now set to grow to fill the remaining space in the parent div. This allows each Product component to be the same size and not look weird.